+++ modules/ext.wikispeech.ui.js
@@ -15,17 +15,23 @@
 		 */
 
 		this.init = function () {
-			mw.user.getRights()
-				.done( function ( rights ) {
-					var canEditLexicon = rights.indexOf( 'wikispeech-edit-lexicon' ) >= 0;
-					mw.wikispeech.ui.addControlPanel( canEditLexicon );
-				} )
-				.fail( function () {
-					// If we can not get the rights we still want to
-					// show the player. Assume the user does not have
-					// the right to edit lexicon.
-					mw.wikispeech.ui.addControlPanel( false );
-				} );
+			if ( mw.wikispeech.consumerMode ) {
+				// No edit lexicon button since we can not check the rights
+				// on the producer.
+				self.addControlPanel( false );
+			} else {
+				mw.user.getRights()
+					.done( function ( rights ) {
+						var canEditLexicon = rights.indexOf( 'wikispeech-edit-lexicon' ) >= 0;
+						mw.wikispeech.ui.addControlPanel( canEditLexicon );
+					} )
+					.fail( function () {
+						// If we can not get the rights we still want to
+						// show the player. Assume the user does not have
+						// the right to edit lexicon.
+						mw.wikispeech.ui.addControlPanel( false );
+					} );
+			}
 			mw.wikispeech.ui.addSelectionPlayer();
 			mw.wikispeech.ui.addBufferingIcon();
 			mw.wikispeech.ui.addKeyboardShortcuts();